+++ spin2/src/spin2.semanticAndOutline.ts
@@ -1,11 +1,13 @@
 "use strict";
-
-import { deepStrictEqual } from "assert";
 // src/spin2.semanticAndOutline.ts
+
+//import { deepStrictEqual } from "assert";
+
+import { semanticConfiguration, reloadSemanticConfiguration } from "./spin2.semantic.configuration";
 
 import * as vscode from "vscode";
 
-// ----------------------------------------------------------------------------
+// ============================================================================
 //  this file contains both an outline provider
 //    and our semantic highlighting provider
 //
@@ -238,9 +240,11 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     // SEE https://www.codota.com/code/javascript/functions/vscode/CancellationToken/isCancellationRequested
     if (cancelToken) {
     } // silence our compiler for now... TODO: we should adjust loop so it can break on cancelToken.isCancellationRequested
+    this._resetForNewDocument();
+    this._logMessage("* Config: spinExtensionBehavior.highlightFlexspinDirectives: [" + this.configuration.highlightFlexspin + "]");
+
     const allTokens = this._parseText(document.getText());
     const builder = new vscode.SemanticTokensBuilder();
-    this._resetForNewDocument();
     allTokens.forEach((token) => {
       builder.push(token.line, token.startCharacter, token.length, this._encodeTokenType(token.ptTokenType), this._encodeTokenModifiers(token.ptTokenModifiers));
     });
@@ -256,6 +260,8 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private debugDisplays = new Map<string, eDebugDisplayType>();
   // list of directives found in file
   private fileDirectives: ISpin2Directive[] = [];
+
+  private configuration = semanticConfiguration;
 
   private currentMethodName: string = "";
 
@@ -278,6 +284,9 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     this.currentMethodName = "";
     this.debugDisplays.clear();
     this.fileDirectives = [];
+    if (reloadSemanticConfiguration()) {
+      this.configuration = semanticConfiguration;
+    }
   }
 
   private _encodeTokenModifiers(strTokenModifiers: string[]): number {
@@ -321,8 +330,9 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     for (let i = 0; i < lines.length; i++) {
       const line = lines[i];
       const trimmedLine = line.trim();
-      const trimmedNonCommentLine = this._getNonCommentLineRemainder(0, line);
+      const trimmedNonCommentLine: string = this._getNonCommentLineRemainder(0, line);
       const sectionStatus = this._isSectionStartLine(line);
+      const lineParts: string[] = trimmedNonCommentLine.split(/[ \t]/);
       if (currState == eParseState.inMultiLineComment) {
         // in multi-line non-doc-comment, hunt for end '}' to exit
         let closingOffset = line.indexOf("}");
@@ -340,6 +350,9 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           currState = priorState;
         }
         //  DO NOTHING Let Syntax highlighting do this
+        continue;
+      } else if (this._isFlexspinPreprocessorDirective(lineParts[0])) {
+        this._getPreProcessor_Declaration(0, i + 1, line);
         continue;
       } else if (sectionStatus.isSectionStart) {
         currState = sectionStatus.inProgressStatus;
@@ -512,6 +525,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
       const line = lines[i];
       const trimmedLine = line.trim();
       const sectionStatus = this._isSectionStartLine(line);
+      const lineParts: string[] = trimmedLine.split(/[ \t]/);
       // TODO: UNDONE add filter which corrects for syntax inability to mark 'comments when more than one "'" present on line!
       //if (trimmedLine.length > 2 && trimmedLine.includes("'")) {
       //    const partialTokenSet: IParsedToken[] = this._possiblyMarkBrokenSingleLineComment(i, 0, line);
@@ -535,6 +549,13 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           currState = priorState;
         }
         //  DO NOTHING Let Syntax highlighting do this
+      } else if (this._isFlexspinPreprocessorDirective(lineParts[0])) {
+        const partialTokenSet: IParsedToken[] = this._reportPreProcessorLine(i, 0, line);
+        partialTokenSet.forEach((newToken) => {
+          this._logPreProc("=> PreProc: " + this._tokenString(newToken, line));
+          tokenSet.push(newToken);
+        });
+        continue;
       } else if (sectionStatus.isSectionStart) {
         currState = sectionStatus.inProgressStatus;
         this._logState("  -- ln:" + (i + 1) + " currState=[" + currState + "]");
@@ -556,7 +577,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           if (line.length > 3) {
             const partialTokenSet: IParsedToken[] = this._reportCON_DeclarationLine(i, 3, line);
             partialTokenSet.forEach((newToken) => {
-              this._logOBJ("=> CON: " + this._tokenString(newToken, line));
+              this._logCON("=> CON: " + this._tokenString(newToken, line));
               tokenSet.push(newToken);
             });
           }
@@ -659,7 +680,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           this._logCON("- process CON line(" + (i + 1) + "):  trimmedLine=[" + trimmedLine + "]");
           const partialTokenSet: IParsedToken[] = this._reportCON_DeclarationLine(i, 0, line);
           partialTokenSet.forEach((newToken) => {
-            this._logOBJ("=> CON: " + this._tokenString(newToken, line));
+            this._logCON("=> CON: " + this._tokenString(newToken, line));
             tokenSet.push(newToken);
           });
         }
@@ -687,7 +708,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
             // process ORG line allowing label to be present
             const partialTokenSet: IParsedToken[] = this._reportDAT_DeclarationLine(i, 0, line);
             partialTokenSet.forEach((newToken) => {
-              this._logOBJ("=> DAT: " + this._tokenString(newToken, line));
+              this._logDAT("=> DAT: " + this._tokenString(newToken, line));
               tokenSet.push(newToken);
             });
 
@@ -697,7 +718,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           } else {
             const partialTokenSet: IParsedToken[] = this._reportDAT_DeclarationLine(i, 0, line);
             partialTokenSet.forEach((newToken) => {
-              this._logOBJ("=> DAT: " + this._tokenString(newToken, line));
+              this._logDAT("=> DAT: " + this._tokenString(newToken, line));
               tokenSet.push(newToken);
             });
           }
@@ -729,7 +750,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           // in DAT sections we end with FIT or just next section
           const partialTokenSet: IParsedToken[] = this._reportDAT_PasmCode(i, 0, line);
           partialTokenSet.forEach((newToken) => {
-            this._logOBJ("=> DAT: " + this._tokenString(newToken, line));
+            this._logPASM("=> DAT: " + this._tokenString(newToken, line));
             tokenSet.push(newToken);
           });
           const lineParts: string[] = trimmedLine.split(/[ \t]/);
@@ -809,6 +830,25 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     }
   }
 
+  private _getPreProcessor_Declaration(startingOffset: number, lineNbr: number, line: string): void {
+    if (this.configuration.highlightFlexspin) {
+      let currentOffset: number = this._skipWhite(line, startingOffset);
+      const nonCommentConstantLine = this._getNonCommentLineRemainder(currentOffset, line);
+      // get line parts - we only care about first one
+      const lineParts: string[] = nonCommentConstantLine.split(/[ \t=]/);
+      this._logPreProc("  - ln:" + lineNbr + " GetPreProcDecl lineParts=[" + lineParts + "]");
+      const directive: string = lineParts[0];
+      const symbolName: string | undefined = lineParts.length > 1 ? lineParts[1] : undefined;
+      if (this._isFlexspinPreprocessorDirective(directive)) {
+        // check a valid preprocessor line for a declaration
+        if (symbolName != undefined && directive.toLowerCase() == "#define") {
+          this._logPreProc("  -- new PreProc Symbol=[" + symbolName + "]");
+          this._setGlobalToken(symbolName, new RememberedToken("variable", ["readonly"]));
+        }
+      }
+    }
+  }
+
   private _getCON_Declaration(startingOffset: number, lineNbr: number, line: string): void {
     // HAVE    DIGIT_NO_VALUE = -2   ' digit value when NOT [0-9]
     //  -or-   _clkfreq = CLK_FREQ   ' set system clock
@@ -819,7 +859,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
       const nonCommentConstantLine = this._getNonCommentLineRemainder(currentOffset, line);
       this._logCON("  - ln:" + lineNbr + " GetCONDecl nonCommentConstantLine=[" + nonCommentConstantLine + "]");
 
-      const haveEnumDeclaration: boolean = nonCommentConstantLine.indexOf("#") != -1;
+      const haveEnumDeclaration: boolean = nonCommentConstantLine.startsWith("#");
       const containsMultiAssignments: boolean = nonCommentConstantLine.indexOf(",") != -1;
       let statements: string[] = [nonCommentConstantLine];
       if (!haveEnumDeclaration && containsMultiAssignments) {
@@ -923,7 +963,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this._isDatStorageType(lineParts[1]) ? true : false;
     if (haveLabel) {
       const labelName: string = lineParts[0];
-      if (!this._isReservedPasmSymbols(labelName) && !labelName.toUpperCase().startsWith("IF_")) {
+      if (!this._isReservedPasmSymbols(labelName) && !labelName.toUpperCase().startsWith("IF_") && !labelName.toUpperCase().startsWith("_RET_")) {
         // org in first column is not label name, nor is if_ conditional
         const labelType: string = isDataDeclarationLine ? "variable" : "label";
         var labelModifiers: string[] = [];
@@ -1067,13 +1107,94 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     }
   }
 
+  private _reportPreProcessorLine(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
+    const tokenSet: IParsedToken[] = [];
+    // skip Past Whitespace
+    let currentOffset: number = this._skipWhite(line, startingOffset);
+    const nonCommentConstantLine = this._getNonCommentLineRemainder(currentOffset, line);
+    // get line parts - we only care about first one
+    const lineParts: string[] = nonCommentConstantLine.split(/[ \t=]/);
+    this._logPreProc("  - ln:" + lineNumber + " reportPreProc lineParts=[" + lineParts + "]");
+    const directive: string = lineParts[0];
+    const symbolName: string | undefined = lineParts.length > 1 ? lineParts[1] : undefined;
+    if (this.configuration.highlightFlexspin) {
+      if (this._isFlexspinPreprocessorDirective(directive)) {
+        // record the directive
+        tokenSet.push({
+          line: lineNumber,
+          startCharacter: 0,
+          length: directive.length,
+          ptTokenType: "keyword",
+          ptTokenModifiers: ["control", "directive"],
+        });
+        const hasSymbol: boolean =
+          directive.toLowerCase() == "#define" ||
+          directive.toLowerCase() == "#ifdef" ||
+          directive.toLowerCase() == "#ifndef" ||
+          directive.toLowerCase() == "#elseifdef" ||
+          directive.toLowerCase() == "#elseifndef";
+        if (hasSymbol && symbolName != undefined) {
+          const nameOffset = line.indexOf(symbolName, currentOffset);
+          this._logPreProc("  -- GLBL symbolName=[" + symbolName + "]");
+          let referenceDetails: RememberedToken | undefined = undefined;
+          if (this._isGlobalToken(symbolName)) {
+            referenceDetails = this._getGlobalToken(symbolName);
+            this._logPreProc("  --  FOUND preProc global " + this._rememberdTokenString(symbolName, referenceDetails));
+          }
+          if (referenceDetails != undefined) {
+            // record a constant declaration!
+            const updatedModificationSet: string[] = directive.toLowerCase() == "#define" ? this._modifiersWith(referenceDetails.modifiers, "declaration") : referenceDetails.modifiers;
+            tokenSet.push({
+              line: lineNumber,
+              startCharacter: nameOffset,
+              length: symbolName.length,
+              ptTokenType: referenceDetails.type,
+              ptTokenModifiers: updatedModificationSet,
+            });
+          } else if (this._isFlexspinReservedWord(symbolName)) {
+            // record a constant reference
+            tokenSet.push({
+              line: lineNumber,
+              startCharacter: nameOffset,
+              length: symbolName.length,
+              ptTokenType: "variable",
+              ptTokenModifiers: ["readonly"],
+            });
+          } else {
+            // record an unknown name
+            tokenSet.push({
+              line: lineNumber,
+              startCharacter: nameOffset,
+              length: symbolName.length,
+              //ptTokenType: "variable",
+              //ptTokenModifiers: ["readonly", "missingDeclaration"],
+              ptTokenType: "comment",
+              ptTokenModifiers: ["line"],
+            });
+          }
+        }
+      }
+    } else {
+      //  DO NOTHING we don't highlight these (flexspin support not enabled)
+      tokenSet.push({
+        line: lineNumber,
+        startCharacter: 0,
+        length: lineParts[0].length,
+        ptTokenType: "macro",
+        ptTokenModifiers: ["directive", "illegalUse"],
+      });
+    }
+
+    return tokenSet;
+  }
+
   private _reportCON_DeclarationLine(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     // skip Past Whitespace
     let currentOffset: number = this._skipWhite(line, startingOffset);
     const nonCommentConstantLine = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
 
-    const haveEnumDeclaration: boolean = nonCommentConstantLine.indexOf("#") != -1;
+    const haveEnumDeclaration: boolean = nonCommentConstantLine.startsWith("#");
     const containsMultiAssignments: boolean = nonCommentConstantLine.indexOf(",") != -1;
     this._logCON("- reportConstant haveEnum=(" + haveEnumDeclaration + "), containsMulti=(" + containsMultiAssignments + "), nonCommentConstantLine=[" + nonCommentConstantLine + "]");
     let statements: string[] = [nonCommentConstantLine];
@@ -2255,7 +2376,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           argumentOffset++;
           minNonLabelParts++;
         }
-        if (lineParts[argumentOffset].toUpperCase().startsWith("IF_")) {
+        if (lineParts[argumentOffset].toUpperCase().startsWith("IF_") || lineParts[argumentOffset].toUpperCase().startsWith("_RET_")) {
           // skip our conditional
           argumentOffset++;
           minNonLabelParts++;
@@ -2324,7 +2445,8 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                   // we don't have name registered so just mark it
                   if (namePart != ".") {
                     // odd special case!
-                    if (!this._isSpinReservedWord(namePart) && !this._isBuiltinReservedWord(namePart) && !this._isDebugMethod(namePart)) {
+                    if (!this._isSpinReservedWord(namePart) && !this._isBuiltinReservedWord(namePart) && !this._isDebugMethod(namePart) && !this._isPasmModczOperand(namePart)) {
+                      // XYZZY
                       this._logPASM("  --  SPIN Pasm MISSING name=[" + namePart + "]");
                       tokenSet.push({
                         line: lineNumber,
@@ -3302,6 +3424,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   // adjust following true/false to show specific parsing debug
   private spin2DebugLogEnabled: boolean = false; // WARNING (REMOVE BEFORE FLIGHT)- change to 'false' - disable before commit
   private showSpinCode: boolean = true;
+  private showPreProc: boolean = true;
   private showCON: boolean = true;
   private showOBJ: boolean = true;
   private showDAT: boolean = true;
@@ -3325,6 +3448,12 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
 
   private _logSPIN(message: string): void {
     if (this.showSpinCode) {
+      this._logMessage(message);
+    }
+  }
+
+  private _logPreProc(message: string): void {
+    if (this.showPreProc) {
       this._logMessage(message);
     }
   }
@@ -3862,7 +3991,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _getLocalToken(tokenName: string): RememberedToken | undefined {
     const desiredToken: RememberedToken | undefined = this.localTokens.get(tokenName.toLowerCase());
     if (desiredToken != undefined) {
-      this._logOBJ("  -- FND-lTOK " + this._rememberdTokenString(tokenName, desiredToken));
+      this._logTokenSet("  -- FND-lTOK " + this._rememberdTokenString(tokenName, desiredToken));
     }
     return desiredToken;
   }
@@ -4519,6 +4648,38 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
       "qcos",
     ];
     const reservedStatus: boolean = spinMethodNames.indexOf(name.toLowerCase()) != -1;
+    return reservedStatus;
+  }
+
+  private _isFlexspinPreprocessorDirective(name: string): boolean {
+    const flexspinDirectiveOfNote: string[] = ["#define", "#ifdef", "#ifndef", "#else", "#elseifdef", "#elseifndef", "#endif", "#error", "#include", "#warn", "#undef"];
+    const reservedStatus: boolean = flexspinDirectiveOfNote.indexOf(name.toLowerCase()) != -1;
+    return reservedStatus;
+  }
+
+  private _isFlexspinReservedWord(name: string): boolean {
+    const flexspinReservedswordsOfNote: string[] = [
+      "__propeller__",
+      "__propeller2__",
+      "__p2__",
+      "__flexspin__",
+      "__spincvt__",
+      "__spin2pasm__",
+      "__spin2cpp__",
+      "__have_fcache__",
+      "__cplusplus__",
+      "__date__",
+      "__file__",
+      "__line__",
+      "__time__",
+      "__version__",
+      "__debug__",
+      "__output_asm__",
+      "__output_bytecode__",
+      "__output_c__",
+      "__output_cpp__",
+    ];
+    const reservedStatus: boolean = flexspinReservedswordsOfNote.indexOf(name.toLowerCase()) != -1;
     return reservedStatus;
   }